requirements.txt

keras==2.1.2
tensorflow==1.4.1
biopython
pybedtools
PyVCF
h5py
numpy
pybedtools
scikit-learn
pandas
seaborn


